Entity 오류 수정 springboot 버전 수정

+++ src/main/java/com/example/membership/service/infra/entity/History.java
@@ -5,9 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.sql.Timestamp;
-
 @Getter
 @Setter
 @AllArgsConstructor
@@ -16,7 +15,7 @@ import java.sql.Timestamp;
 public class History {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "point_id")
     private Point point_id;
